+++ inc/Messages/ScalarMessage.h
@@ -41,7 +41,7 @@ namespace PCOE {
         }
 
     protected:
-        std::uint16_t getPayloadSize() const override final {
+        std::uint32_t getPayloadSize() const override final {
             return sizeof(T);
         }
