+++ inc/Messages/WaypointMessage.h
@@ -72,7 +72,7 @@ namespace PCOE {
         }
 
     protected:
-        std::uint16_t getPayloadSize() const override final;
+        std::uint32_t getPayloadSize() const override final;
 
         void serializePayload(std::ostream& os) const override final;
